add ignored users hook
